+++ sdk/src/dlob/DLOBSubscriber.ts
@@ -16,7 +16,6 @@ import {
 	L2OrderBookGenerator,
 	L3OrderBook,
 } from './orderBookLevels';
-import { calculateAskPrice, calculateBidPrice } from '../math/market';
 
 export class DLOBSubscriber {
 	driftClient: DriftClient;
@@ -103,15 +102,7 @@ export class DLOBSubscriber {
 			}
 		}
 
-		if (includeVamm && fallbackL2Generators.length > 0) {
-			throw new Error(
-				'includeVamm can only be used if fallbackL2Generators is empty'
-			);
-		}
-
 		let oraclePriceData;
-		let fallbackBid;
-		let fallbackAsk;
 		const isPerp = isVariant(marketType, 'perp');
 		if (isPerp) {
 			const perpMarketAccount =
@@ -119,14 +110,18 @@ export class DLOBSubscriber {
 			oraclePriceData = this.driftClient.getOracleDataForPerpMarket(
 				perpMarketAccount.marketIndex
 			);
-			fallbackBid = calculateBidPrice(perpMarketAccount, oraclePriceData);
-			fallbackAsk = calculateAskPrice(perpMarketAccount, oraclePriceData);
 		} else {
 			oraclePriceData =
 				this.driftClient.getOracleDataForSpotMarket(marketIndex);
 		}
 
 		if (isPerp && includeVamm) {
+			if (fallbackL2Generators.length > 0) {
+				throw new Error(
+					'includeVamm can only be used if fallbackL2Generators is empty'
+				);
+			}
+
 			fallbackL2Generators = [
 				getVammL2Generator({
 					marketAccount: this.driftClient.getPerpMarketAccount(marketIndex),
@@ -143,8 +138,6 @@ export class DLOBSubscriber {
 			depth,
 			oraclePriceData,
 			slot: this.slotSource.getSlot(),
-			fallbackBid,
-			fallbackAsk,
 			fallbackL2Generators: fallbackL2Generators,
 		});
 	}